Relocate ForeignKeyDbTransformer
