+++ src/schemaType/BaseSchemaType.ts
@@ -1,4 +1,4 @@
-import type { ForeignKeyDbDefinition } from '../shared/classes/ForeignKeyDbTransformer';
+import type { ForeignKeyDbDefinition } from '../ForeignKeyDbTransformer';
 import type { GenericObject, MvRecord, Validator } from '../shared/types';
 
 export interface SchemaTypeDefinitionBase {
