+++ src/schemaType/DocumentArrayType.ts
@@ -1,7 +1,7 @@
 import { cloneDeep, isPlainObject, set as setIn } from 'lodash';
 import Document from '../Document';
+import type { ForeignKeyDbDefinition } from '../ForeignKeyDbTransformer';
 import type Schema from '../Schema';
-import type { ForeignKeyDbDefinition } from '../shared/classes/ForeignKeyDbTransformer';
 import type { MvRecord } from '../shared/types';
 import { ensureArray, getFromMvArray } from '../utils';
 import BaseSchemaType from './BaseSchemaType';
