+++ src/schemaType/__tests__/ArrayType.test.ts
@@ -1,5 +1,5 @@
 import { NumberType, StringType } from '..';
-import type { ForeignKeyDbDefinition } from '../../shared/classes/ForeignKeyDbTransformer';
+import type { ForeignKeyDbDefinition } from '../../ForeignKeyDbTransformer';
 import type { MvRecord } from '../../shared/types';
 import ArrayType from '../ArrayType';
 import type { SchemaTypeDefinitionNumber } from '../NumberType';
