+++ src/schemaType/__tests__/NestedArrayType.test.ts
@@ -1,5 +1,5 @@
 import { NumberType, StringType } from '..';
-import type { ForeignKeyDbDefinition } from '../../shared/classes/ForeignKeyDbTransformer';
+import type { ForeignKeyDbDefinition } from '../../ForeignKeyDbTransformer';
 import type { MvRecord } from '../../shared/types';
 import NestedArrayType from '../NestedArrayType';
 import type { SchemaTypeDefinitionNumber } from '../NumberType';
