--- src/shared/classes/index.ts
@@ -1,1 +1,0 @@
-export { default as ForeignKeyDbTransformer } from './ForeignKeyDbTransformer';
